Move some common code to common.go.
